+++ packages/core/src/actions/sendTransaction.test.ts
@@ -75,10 +75,9 @@ test('behavior: value exceeds balance', async () => {
       from:   0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266
       to:     0xd2135CfB216b74109775236E36d4b433F1DF507B
       value:  99999 ETH
-      gas:    21000
 
     Details: Insufficient funds for gas * price + value
-    Version: viem@2.9.25"
+    Version: viem@2.9.31"
   `)
   await disconnect(config, { connector })
 })