+++ packages/core/src/actions/signMessage.test.ts
@@ -51,7 +51,7 @@ test('behavior: user rejected request', async () => {
     [UserRejectedRequestError: User rejected the request.
 
     Details: Failed to sign message.
-    Version: viem@2.9.25]
+    Version: viem@2.9.31]
   `)
   await disconnect(config, { connector: connector_ })
 })