+++ packages/core/src/actions/waitForTransactionReceipt.ts
@@ -63,15 +63,15 @@ export async function waitForTransactionReceipt<
     )
     const txn = await action_getTransaction({ hash: receipt.transactionHash })
     const action_call = getAction(client, call, 'call')
-    const code = (await action_call({
+    const code = await action_call({
       ...(txn as any),
       gasPrice: txn.type !== 'eip1559' ? txn.gasPrice : undefined,
       maxFeePerGas: txn.type === 'eip1559' ? txn.maxFeePerGas : undefined,
       maxPriorityFeePerGas:
         txn.type === 'eip1559' ? txn.maxPriorityFeePerGas : undefined,
-    })) as unknown as string
-    const reason = code
-      ? hexToString(`0x${code.substring(138)}`)
+    })
+    const reason = code?.data
+      ? hexToString(`0x${code.data.substring(138)}`)
       : 'unknown reason'
     throw new Error(reason)
   }