+++ packages/core/src/experimental/query/writeContracts.test.ts
@@ -1,0 +1,15 @@
+import { config } from '@wagmi/test'
+import { expect, test } from 'vitest'
+
+import { writeContractsMutationOptions } from './writeContracts.js'
+
+test('default', () => {
+  expect(writeContractsMutationOptions(config)).toMatchInlineSnapshot(`
+    {
+      "mutationFn": [Function],
+      "mutationKey": [
+        "writeContracts",
+      ],
+    }
+  `)
+})