+++ playgrounds/vite-core/src/wagmi.ts
@@ -1,4 +1,4 @@
-import { coinbaseWallet, walletConnect } from '@wagmi/connectors'
+import { coinbaseWallet, metaMask, walletConnect } from '@wagmi/connectors'
 import { http, createConfig, createStorage } from '@wagmi/core'
 import { mainnet, optimism, sepolia } from '@wagmi/core/chains'
 
@@ -7,6 +7,7 @@ export const config = createConfig({
   connectors: [
     walletConnect({ projectId: import.meta.env.VITE_WC_PROJECT_ID }),
     coinbaseWallet({ appName: 'Vite React Playground' }),
+    metaMask(),
   ],
   storage: createStorage({ storage: localStorage, key: 'vite-core' }),
   transports: {